+++ src/core/ctx.rs
@@ -1,10 +1,13 @@
-use bevy::{prelude::*, reflect::TypeRegistry};
+use bevy::{ecs::component::ComponentId, prelude::*, reflect::TypeRegistry};
 
 use super::{replicon_tick::RepliconTick, server_entity_map::ServerEntityMap, Replicated};
 
 /// Replication context for serialization function.
 #[non_exhaustive]
 pub struct SerializeCtx {
+    /// ID of the serializing component.
+    pub component_id: ComponentId,
+
     /// Current tick.
     pub server_tick: RepliconTick,
 }
@@ -18,6 +21,9 @@ pub struct WriteCtx<'a, 'w, 's> {
     /// Maps server entities to client entities and vice versa.
     pub entity_map: &'a mut ServerEntityMap,
 
+    /// ID of the writing component.
+    pub component_id: ComponentId,
+
     /// Tick for the currently processing message.
     pub message_tick: RepliconTick,
 
@@ -29,11 +35,13 @@ impl<'a, 'w, 's> WriteCtx<'a, 'w, 's> {
     pub(crate) fn new(
         commands: &'a mut Commands<'w, 's>,
         entity_map: &'a mut ServerEntityMap,
+        component_id: ComponentId,
         message_tick: RepliconTick,
     ) -> Self {
         Self {
             commands,
             entity_map,
+            component_id,
             message_tick,
             ignore_mapping: false,
         }
@@ -57,17 +65,11 @@ pub struct RemoveCtx<'a, 'w, 's> {
     /// A queue to perform structural changes to the [`World`].
     pub commands: &'a mut Commands<'w, 's>,
 
+    /// ID of the removing component.
+    pub component_id: ComponentId,
+
     /// Tick for the currently processing message.
     pub message_tick: RepliconTick,
-}
-
-impl<'a, 'w, 's> RemoveCtx<'a, 'w, 's> {
-    pub(crate) fn new(commands: &'a mut Commands<'w, 's>, message_tick: RepliconTick) -> Self {
-        Self {
-            commands,
-            message_tick,
-        }
-    }
 }
 
 /// Replication context for despawn.