+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.28.4] - 2024-10-15
 
+### Added
+
+- `ComponentId` to all component-related contexts.
+
+### Changed
+
+- All custom functions now accept the newly added `DeferredEntity` instead of `EntityMut`. This new entity type provides read-only access to the world.
+
 ### Fixed
 
 - Synchronize server events with init messages properly when `ServerTick` is not updated every app tick.